refactor generatePlotlyProps and generateDefaultChartAxes
