+++ libs/counterfactuals/.eslintrc.json
@@ -16,7 +16,9 @@
     },
     {
       "files": ["*.ts", "*.tsx"],
-      "rules": {}
+      "rules": {
+        "max-lines": "error"
+      }
     },
     {
       "files": ["*.js", "*.jsx"],
